requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.7
django-heroku==0.3.1
django-phonenumber-field==5.0.0
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
phonenumbers==8.12.20
progress==1.5
psycopg2==2.8.6
PyJWT==1.7.1
pyparsing==2.4.6
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
twilio==6.54.0
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
